Change payment_partnerCost type to Double

+++ carma-models/src/Carma/Model/Service.hs
@@ -31,7 +31,7 @@ data Service = Service
                                  "Тип оплаты"
   , payment_costTranscript       :: F (Maybe Text) "payment_costTranscript"
                                  "Расшифровка стоимости"
-  , payment_partnerCost          :: F (Maybe Int) "payment_partnerCost"
+  , payment_partnerCost          :: F (Maybe Double) "payment_partnerCost"
                                  "Стоимость со слов партнёра (число)"
   , payment_calculatedCost       :: F (Maybe Int) "payment_calculatedCost"
                                  "Расчётная стоимость"
